feat/DirrochaCMS-21: refatorando, componentizando, normalizando o projeto
