+++ components/footer/index.tsx
@@ -1,0 +1,24 @@
+import Image from "next/image";
+
+export function FooterComponent(){
+    return <>
+    <footer className="row-start-3 flex gap-6 flex-wrap items-center justify-center">
+        <a
+          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
+          href="https://github.com/marco0antonio0/DirrochaCMS"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          <Image
+            aria-hidden
+            src="https://nextjs.org/icons/file.svg"
+            alt="File icon"
+            width={16}
+            height={16}
+          />
+          Desenvolvido por @marco0antonio0
+        </a>
+        
+      </footer>
+    </>
+}
