+++ utils/formatDataToDynamicObject.ts
@@ -1,3 +1,5 @@
+import { typeFormat } from "./typesFormat";
+
 export const formatDataToDynamicObject = (data: any) => {
     // Se `data` for um objeto único, convertemos para um array
     if (!Array.isArray(data)) {
@@ -5,20 +7,7 @@ export const formatDataToDynamicObject = (data: any) => {
         data = [data]; // Transforma em um array
     }
 
-    const fieldProperties: { [key: string]: { mult: boolean; type?: string } } = {
-        titulo_identificador: { mult: false, type: "string" },
-        data: { mult: false, type: "date" },
-        link: { mult: false, type: "string" },
-        preco: { mult: false, type: "string" },
-        titulo: { mult: false, type: "string" },
-        descricao: { mult: true, type: "string" },
-        breve_descricao: { mult: true, type: "string" },
-        artigo: { mult: true, type: "string" },
-        image: { mult: false, type: "img" },
-        nome: { mult: false, type: "string" },
-        senha: { mult: false, type: "string" },
-        texto: { mult: true, type: "string" },
-    };
+    const fieldProperties: { [key: string]: { mult: boolean; type?: string } } = typeFormat
 
     return data.map((item:any) => ({
         id_endpoint: item.endpointId || null,
