+++ utils/typesFormat.ts
@@ -1,0 +1,14 @@
+export const typeFormat = {
+    titulo_identificador: { mult: false, type: "string" },
+    data:           { mult: false, type: "date" },
+    link:           { mult: false, type: "string" },
+    preco:          { mult: false, type: "string" },
+    titulo:         { mult: false, type: "string" },
+    descricao:      { mult: true,  type: "string" },
+    breve_descricao:{ mult: true,  type: "string" },
+    artigo:         { mult: true,  type: "string" },
+    image:          { mult: false, type: "img" },
+    nome:           { mult: false, type: "string" },
+    senha:          { mult: false, type: "string" },
+    texto:          { mult: true,  type: "string" },
+};
